+++ agent/api/taskstatus_test.go
@@ -1,0 +1,44 @@
+// Copyright 2014-2017 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License"). You may
+// not use this file except in compliance with the License. A copy of the
+// License is located at
+//
+//	http://aws.amazon.com/apache2.0/
+//
+// or in the "license" file accompanying this file. This file is distributed
+// on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+// express or implied. See the License for the specific language governing
+// permissions and limitations under the License.
+
+package api
+
+import (
+	"encoding/json"
+	"testing"
+)
+
+type testTaskStatus struct {
+	SomeStatus TaskStatus `json:"status"`
+}
+
+func TestUnmarshalTaskStatus(t *testing.T) {
+	status := TaskStatusNone
+
+	err := json.Unmarshal([]byte(`"RUNNING"`), &status)
+	if err != nil {
+		t.Error(err)
+	}
+	if status != TaskRunning {
+		t.Error("RUNNING should unmarshal to RUNNING, not " + status.String())
+	}
+
+	var test testTaskStatus
+	err = json.Unmarshal([]byte(`{"status":"STOPPED"}`), &test)
+	if err != nil {
+		t.Error(err)
+	}
+	if test.SomeStatus != TaskStopped {
+		t.Error("STOPPED should unmarshal to STOPPED, not " + test.SomeStatus.String())
+	}
+}
